feat: update all the folders to no longer appear under users
