+++ app/models/gdpr_delete_request.rb
@@ -1,0 +1,7 @@
+class GdprDeleteRequest < ApplicationRecord
+  validates :email, :user_id, presence: true
+
+  def self.table_name_prefix
+    "users_"
+  end
+end
