--- app/models/users/gdpr_delete_request.rb
@@ -1,9 +1,0 @@
-module Users
-  class GdprDeleteRequest < ApplicationRecord
-    validates :email, :user_id, presence: true
-
-    def self.table_name_prefix
-      "users_"
-    end
-  end
-end
