+++ app/views/admin/gdpr_delete_requests/index.html.erb
